requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
cx_Oracle==8.3.0
daphne==4.1.2
defusedxml==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-crontab==0.7.1
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-extensions==3.2.3
django-filter==24.1
django-nine==0.2.7
django-rest-auth==0.9.5
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
elastic-transport==8.13.0
elasticsearch==8.13.0
elasticsearch-dsl==8.13.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
kombu==5.3.6
lxml==5.2.1
Markdown==3.5.2
msgpack==1.0.8
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.0
pika==1.3.2
pillow==10.3.0
pluggy==1.5.0
prompt-toolkit==3.0.43
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodome==3.20.0
pydotplus==2.0.2
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
pytesseract==0.3.10
pytest==8.2.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tk==0.1.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
zope.interface==6.4.post2
